Rename file throughout repo
